+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 
+- csharp-mvc: restore by url
 - csharp: restore by url
 - nodejs: restore by url
 - nodejs: support vsdx in diagram editor